requirements.txt
django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1